% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptMappings.R
\name{addDot}
\alias{addDot}
\title{Add Dot in ICD10 Codes}
\usage{
addDot(input_string)
}
\arguments{
\item{input_string}{An ICD10 string.}
}
\value{
A string with a dot added at the appropriate position.
}
\description{
This function adds a dot in ICD10 codes at the appropriate position.
}
\examples{
\dontrun{
  icd10_with_dot <- addDot("A001")
}
}
