% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CirceR2CDMConnector.R
\name{CirceR2CDMConn}
\alias{CirceR2CDMConn}
\title{Create Cohort Set from Named Cohort List}
\usage{
CirceR2CDMConn(named_cohort_list)
}
\arguments{
\item{named_cohort_list}{A named list of CirceR cohorts.}
}
\value{
A \code{CohortSet} s3 object containing the cohort definitions.
}
\description{
This function creates a cohort set from a named list of cohorts and returns it as a \code{CohortSet} object.
}
\examples{
\dontrun{
cdm <- tidyOhdsiRecipies::tidyCdmMock()
caprCs1 <- Capr::cs(4133224, name = "lobar pneumonia")
caprCs2 <- Capr::cs(45176377, name = "velcade")

csWithDetails <- purrr::map(
  list(caprCs1, caprCs2), ~
    tidyOhdsiRecipies::getCaprCsDetails(.x, cdm)
)
cohortsBasedOnCs <- purrr::map(
  csWithDetails, ~ tidyOhdsiRecipies::createCaprConceptSetCohort(.x, addSourceCriteria = TRUE)
)

named_cohort_list <- list(
  Lobar_pneumonia = cohortsBasedOnCs[[1]],
  velcade = cohortsBasedOnCs[[2]]
)

cohortSet <- CirceR2CDMConn(named_cohort_list)
}
}
