% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CirceR2CDMConnector.R
\name{cohortsToCreate2CDMConn}
\alias{cohortsToCreate2CDMConn}
\title{Convert Cohorts To Create to CDM Connector Format}
\usage{
cohortsToCreate2CDMConn(cohortsToCreate)
}
\arguments{
\item{cohortsToCreate}{A data frame containing the cohorts to be converted. The data frame should include columns for cohort ID, cohort name, JSON, and SQL.}
}
\value{
A data frame in the CDM Connector format with class "CohortSet".
}
\description{
This function converts a data frame of cohorts to a format compatible with the CDM Connector.
}
\examples{
library(tidyOhdsiRecipies)
path <- fs::path(fs::path_package("tidyOhdsiRecipies"), "cohorts")
cohortsToCreate <- tidyOhdsiRecipies::createCohortsToCreate(path)
CohortSet <- cohortsToCreate2CDMConn(cohortsToCreate)
}
