% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectCandidates.R
\name{collectCandidatesToCapr}
\alias{collectCandidatesToCapr}
\title{Function Collects Reg-ex Matches of Chosen Domain}
\usage{
collectCandidatesToCapr(
  con,
  vocabularyDatabaseSchema,
  keywords,
  exclude = NULL,
  domains = "Condition"
)
}
\arguments{
\item{con}{A connection to an OMOP CDM database}

\item{vocabularyDatabaseSchema}{Schema name where your OMOP vocabulary format resides. Note that
for SQL Server, this should include both the database and schema
name, for example 'vocabulary.dbo'.}

\item{keywords}{case independent  string pattern to search}

\item{exclude}{case independent  string pattern to exclude}

\item{domains}{case independent  domain_id to search}
}
\value{
Capr concept set
}
\description{
Function Collects Reg-ex Matches of Chosen Domain
}
\examples{
\dontrun{
con <- DatabaseConnector::connect(connectionDetails)
caprCand <- tidyOhdsiRecipies::collectCandidatesToCapr(
  con, "cdm5", c("lobar")
)
}
}
