% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useCohortConceptSets.R
\name{collectConceptIdsFromCohorts}
\alias{collectConceptIdsFromCohorts}
\title{Returns Named List Concept Ids From a Cohort}
\usage{
collectConceptIdsFromCohorts(cohort, con, vocabularyDatabaseSchema)
}
\arguments{
\item{cohort}{A list containing a cohort \code{jsonlite::read_json} or \code{CirceR} output definition}

\item{con}{DBI or DatabaseConnector connection to database}

\item{vocabularyDatabaseSchema}{schema with CDM Vocabulary tables}
}
\value{
named list of numeric vectors
}
\description{
Returns Named List Concept Ids From a Cohort
}
\examples{
\dontrun{
cIds <- collectConceptIdsFromCohorts(cohortDonor, con, 'cdm5')
}
}
