% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptSetDetails.R
\name{collectIngredientConceptIds}
\alias{collectIngredientConceptIds}
\title{Collects Concept Sets or Concept Ids from ingredients}
\usage{
collectIngredientConceptIds(
  ingredients,
  con,
  vocabularyDatabaseSchema,
  return = c("Capr", "descendantIds")
)
}
\arguments{
\item{ingredients}{an ingredient character vector}

\item{con}{DBI or DatabaseConnector connection to database}

\item{vocabularyDatabaseSchema}{CDM schema with vocabulary tables}

\item{return}{define what to return list of Capr or list of concept ids}
}
\value{
either named list of Capr concept sets or named list of numeric vectors with descendant concept ids
}
\description{
Collects Concept Sets or Concept Ids from ingredients
}
\examples{
\dontrun{
cs <- collectIngredientConceptIds(
'vedolizumab', con, 'cdm5', 'Capr')
}
}
