% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptMappings.R
\name{conceptIdsFromSources}
\alias{conceptIdsFromSources}
\title{Get Concept Ids From Concept Codes}
\usage{
conceptIdsFromSources(cdm, listCodes, vocabularyIds = c("ICD10CM"))
}
\arguments{
\item{cdm}{A cdm interface.}

\item{listCodes}{A vector of concept codes.}

\item{vocabularyIds}{A vector of vocabularies. Default is 'ICD10CM'.}
}
\value{
A vector of concept IDs.
}
\description{
This function retrieves concept IDs from a given list of concept codes and vocabularies.
}
\examples{
\dontrun{
  conceptIds <- conceptIdsFromSources(cdm, c("A01", "B02"), c("ICD10CM"))
}
}
