% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptSetTransition.R
\name{conceptSetExpression2CaprCs}
\alias{conceptSetExpression2CaprCs}
\title{Convert Concept Set Expression to Capr Concept Set}
\usage{
conceptSetExpression2CaprCs(x)
}
\arguments{
\item{x}{A named list containing tibbles of concept set expression details.}
}
\value{
A Capr concept set.
}
\description{
This function converts a named list containing tibbles of concept set expression details into a Capr concept set.
}
\examples{
# Example usage:
caprCs <- conceptSetExpression2CaprCs(
  list(test_set = dplyr::tibble(
    concept_id = c(1, 2),
    excluded = c(FALSE, TRUE),
    descendants = c(TRUE, FALSE),
    mapped = c(FALSE, TRUE)
  ))
)
}
