% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyConceptBasedCohort.R
\name{createCaprConceptSetCohort}
\alias{createCaprConceptSetCohort}
\title{Create Capr Concept Set Based Cohort}
\usage{
createCaprConceptSetCohort(
  conceptSet,
  limit = "first",
  requiredObservation = c(0, 0),
  end = "observation_period_end_date",
  endArgs = list(conceptSet = conceptSet, persistenceWindow = 30L, surveillanceWindow =
    0L, daysSupplyOverride = NULL, index = c("startDate"), offsetDays = 7),
  addSourceCriteria = FALSE
)
}
\arguments{
\item{conceptSet}{Capr concept set (Capr S4 class) with details.}

\item{limit}{A character string specifying the limit of events. Possible values are "first", "all", and "last".}

\item{requiredObservation}{A vector of two integers specifying the observation period before and after the index date.}

\item{end}{A character string specifying the end strategy. Possible values are "observation_period_end_date", "drug_exit", and "fixed_exit".}

\item{endArgs}{A list of parameters that define the end strategy for the cohort. The list can include:
\itemize{
\item \code{conceptSet}: The Capr concept set used to define the cohort.
\item \code{persistenceWindow}: An integer specifying the number of days after the index event during which subsequent events are considered part of the same episode.
\item \code{surveillanceWindow}: An integer specifying the number of days after the persistence window during which no events should occur for the episode to be considered ended.
\item \code{daysSupplyOverride}: An optional integer to override the days supply for drug exposures.
\item \code{index}: A character vector specifying the index date(s) for the cohort. Possible values include "startDate" and "endDate".
\item \code{offsetDays}: An integer specifying the number of days to offset the end date from the index date.
}}

\item{addSourceCriteria}{A boolean flag indicating whether to include source criteria in the cohort definition.
\itemize{
\item \code{TRUE}: Source criteria will be added, which may include additional conditions or attributes to refine the cohort.
\item \code{FALSE}: Source criteria will not be added, and the cohort will be defined solely based on the primary criteria.
}}
}
\value{
A CirceR cohort.
}
\description{
This function creates a cohort based on a Capr concept set with detailed information.
}
\examples{
\dontrun{
caprCs <- getCaprCsDetails(
  Capr::cs(list(123456, 789012), name = "Example Concept Set"), cdm
)

coh <- createCaprConceptSetCohort(caprCs)
}
}
