% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyCohortsToCreate.R
\name{createCohortsToCreate}
\alias{createCohortsToCreate}
\title{Create Cohorts To Create Tibble}
\usage{
createCohortsToCreate(path = NULL, cohorts = NULL, computeAttrition = TRUE)
}
\arguments{
\item{path}{A character string specifying the path to the cohort JSON files.}

\item{cohorts}{A named list of \code{CirceR} cohorts.}

\item{computeAttrition}{A boolean indicating whether to compute attrition statistics.}
}
\value{
A tibble for the \code{CohortGenerator} package.
}
\description{
This function creates a tibble for the \code{CohortGenerator} package from cohort JSON files or a list of \code{CirceR} cohorts.
}
\examples{
\dontrun{
ff <- list.files("inst/concept_sets",
  pattern = "cs_",
  full.names = TRUE
)
conceptSetList <- map(ff, ~ Capr::readConceptSet(
  .x,
  name = fs::path_file(.x) |>
    fs::path_ext_remove()
))

cohorts <- map(conceptSetList, ~ createCaprCohort(.x)) |>
  rlang::set_names(
    ff |> fs::path_file() |>
      fs::path_ext_remove()
  )
}
}
