% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptSetDetails.R
\name{getCaprCsDetails}
\alias{getCaprCsDetails}
\title{Get Capr Concept Set Details from CDM}
\usage{
getCaprCsDetails(x, cdm)
}
\arguments{
\item{x}{A list of concept IDs or a Capr concept set (Capr S4 class).}

\item{cdm}{An interface to the CDM, typically a database connection or similar object.}
}
\value{
A Capr concept set with detailed information populated from the CDM.
}
\description{
This function retrieves detailed information for a given Capr concept set or a list of concept IDs from the Common Data Model (CDM).
}
\examples{
\dontrun{
  # Example with a vector of concept IDs
  concept_ids <- c(123456, 789012)
  cdm <- CDMConnector::cdmFromCon(con, 'cdm_schema', 'work_schema')
  detailed_concept_set <- getCaprCsDetails(concept_ids, cdm)

  # Example with a Capr concept set
  capr_concept_set <- Capr::cs(concept_ids, name = "Example Concept Set")
  detailed_concept_set <- getCaprCsDetails(capr_concept_set, cdm)
}
}
