% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptMappings.R
\name{getDescendants}
\alias{getDescendants}
\title{Get Descendant Concepts}
\usage{
getDescendants(x, con, vocabularyDatabaseSchema)
}
\arguments{
\item{x}{A vector of ancestor concept IDs.}

\item{con}{A valid database connection object.}

\item{vocabularyDatabaseSchema}{A character string specifying the schema of the vocabulary database.}
}
\value{
A vector of descendant concept IDs, including the original ancestor concept IDs.
}
\description{
This function retrieves the descendant concept IDs for a given set of ancestor concept IDs from a specified vocabulary database schema.
}
\examples{
\dontrun{
con <- DBI::dbConnect(RPostgres::Postgres(), dbname = "your_database")
ancestor_ids <- c(21164796, 21164797)
vocabulary_schema <- "cdm5"
getDescendants(ancestor_ids, con, vocabulary_schema)
}
}
