% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useCohortConceptSets.R
\name{injectItemsIntoCohort}
\alias{injectItemsIntoCohort}
\title{Inject Items into a Cohort}
\usage{
injectItemsIntoCohort(cohort, caprCs, position, writeCohortPath = NULL)
}
\arguments{
\item{cohort}{A list representing the cohort, which contains ConceptSets.}

\item{caprCs}{An object of class \code{ConceptSet} containing the items to be injected.}

\item{position}{A numeric value indicating the position within the cohort's ConceptSets where the items should be injected. Must be between 1 and the length of the ConceptSets.}

\item{writeCohortPath}{An optional character string specifying the file path to write the modified cohort as a JSON file. Must end with \code{.json}.}
}
\value{
The modified cohort with the injected items.
}
\description{
This function injects items from a given ConceptSet into a specified position within a cohort's ConceptSets.
}
\examples{
# Example usage:
cohortDonor <- jsonlite::read_json(fs::path(
fs::path_package("tidyOhdsiRecipies"), "cohorts", "PHN.json"
))
caprConceptSets <- tidyOhdsiRecipies::collectCaprCsFromCohort(cohortDonor)[[1]]
modifiedCohort <- injectItemsIntoCohort(cohortDonor, caprConceptSets, 2)

}
