% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptSetTransition.R
\name{mergeCsAndConcepts}
\alias{mergeCsAndConcepts}
\title{Union Capr Concept Sets and Concept Ids}
\usage{
mergeCsAndConcepts(x, conceptIds)
}
\arguments{
\item{x}{A named list of \code{ConceptSet}}

\item{conceptIds}{A vector of concept IDs to be included in the merged concept set.}
}
\value{
A \code{ConceptSet} object containing the merged concepts.
}
\description{
This function merges concept sets and concepts into a single concept set.
}
\examples{
\dontrun{
caprConceptSets <- tidyOhdsiRecipies::collectCaprCsFromCohort(cohortDonor)[1]
mergedConceptSet <- mergeCsAndConcepts(caprConceptSets, conceptIds = 1:2)
}
}
