% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareCall.R
\name{prepareCall}
\alias{prepareCall}
\title{Prepare Call for Function in Environment}
\usage{
prepareCall(fn_name, args)
}
\arguments{
\item{fn_name}{A character string specifying the name of the function.}

\item{args}{A named list of arguments to match with the function's parameters.}
}
\value{
A call object for the specified function.
}
\description{
This function prepares a call for a specified function with the given arguments.
}
\examples{
\dontrun{
con <- eval(prepareCall(
  "DatabaseConnector::connect", list(
    connectionDetails = connectionDetails
  )
))
}
}
