% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useCohortConceptSets.R
\name{returnTestDonorCohort}
\alias{returnTestDonorCohort}
\title{Function For Testing}
\usage{
returnTestDonorCohort()
}
\value{
json spec list of cohort definition
}
\description{
Function For Testing
}
\examples{
\dontrun{
cd <- returnTestDonorCohort()
}
}
