% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptSetDetails.R
\name{tidyCaprConceptSetDetails}
\alias{tidyCaprConceptSetDetails}
\title{Tidy Fill in Concept Set details using a vocab}
\usage{
tidyCaprConceptSetDetails(x, con, vocabularyDatabaseSchema = NULL)
}
\arguments{
\item{x}{A concept set created by \code{cs}}

\item{con}{A connection to an OMOP CDM database}

\item{vocabularyDatabaseSchema}{Schema name where your OMOP vocabulary format resides.}
}
\value{
A modified version of the input concept set with concept details filled in.
}
\description{
Concept sets created in R using the \code{cs} function do not contain details like
"CONCEPT_NAME", "DOMAIN_ID", etc. If an OMOP CDM vocabulary is available then
these details can be filled in by the the \code{getConceptSetDetails} function.
Grabed from \verb{https://github.com/OHDSI/Capr} the only difference - added source spec
}
\examples{
\dontrun{
# create a concept set
vocabularyDatabaseSchema <- "main"
Pneumonia <- cs(4133224, name = "Pneumonia")

# fill in the details from an OMOP CDM
con <- returnSqLiteDatabaseConnectorCon()
Pneumonia <- getConceptSetDetails(Pneumonia, con, vocabularyDatabaseSchema = "cdm5")
}
}
