% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyDarwin.R
\name{tidyCdmFromCon}
\alias{tidyCdmFromCon}
\title{Tidy CDM Reference From DBI Con}
\usage{
tidyCdmFromCon(
  con,
  cdmSchema,
  writeSchema,
  cohortTables = NULL,
  cdmName = "db_id"
)
}
\arguments{
\item{con}{A database connection DBI object.}

\item{cdmSchema}{The schema name where the CDM tables are located.}

\item{writeSchema}{The schema name where the cohort tables are located.}

\item{cohortTables}{A character vector of cohort table names. Default is NULL.}

\item{cdmName}{The name of the CDM. Default is the same as \code{cdmSchema}.}
}
\value{
A list of tidied CDM tables.
}
\description{
This function collects and tidies OMOP CDM tables from a given database connection and schema.
}
\examples{
\dontrun{
con <- DBI::dbConnect(RPostgres::Postgres(), dbname = "your_database")
cdmSchema <- "cdm_schema"
writeSchema <- "write_schema"
cohortTables <- c("cohort1", "cohort2")
cdm <- tidyCdm(con, cdmSchema, writeSchema, cohortTables)
}
}
