% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyOhdsiMock.R
\name{tidyCdmMock}
\alias{tidyCdmMock}
\title{Create Mock CDM Reference}
\usage{
tidyCdmMock()
}
\value{
cdm_reference via CDMConnector.
}
\description{
This function creates mock CDM data reference for testing purposes.
}
\examples{
\dontrun{
mock_cdm <- tidyCdmMock()
}
}
