% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Caprstructor.R
\name{writeCaprstructor}
\alias{writeCaprstructor}
\title{Write Capr Cohort Using Constructor}
\usage{
writeCaprstructor(
  entryDomains = c("drug", "condition"),
  firstTime = FALSE,
  requiredObservation = c(1, 1),
  attritionDomains = c("drug", "condition"),
  limit = "first",
  age = 18,
  gender = "male",
  exitStrategy = "observation_period_end_date",
  copyToClipboard = FALSE
)
}
\arguments{
\item{entryDomains}{A character vector specifying the domains for entry events. Default is c('drug', 'condition').}

\item{firstTime}{A logical value indicating if the entry event is the first time. Default is FALSE.}

\item{requiredObservation}{A vector of two integers specifying the observation period before and after the index date.}

\item{attritionDomains}{A character vector specifying the domains for attrition events. Default is c('drug', 'condition'). Might be NULL if no attrition required}

\item{limit}{A character string specifying the limit of events. Possible values are "first", "all", and "last".}

\item{age}{An integer specifying the age for the entry event and it's defined as greater or equal the age}

\item{gender}{A character string specifying the gender for the entry event. Default is "male".}

\item{exitStrategy}{A character string specifying the end strategy. Possible values are "observation_period_end_date", "drug_exit" and "fixed_exit".}

\item{copyToClipboard}{boolean, if TRUE will copy  to clipboard cohort definition}
}
\value{
The function returns a Capr cohort vertical data class
}
\description{
The function allows you to build a template Capr
cohort using a set of arguments.
Most often, difficulties with using Capr arise
due to a lack of understanding of the
structure and the problem of “where to start”
Therefore, \code{writeCaprstructor} allows you to get a
structure with the necessary fields.
}
\examples{
\dontrun{
writeCaprstructor(
  entryDomains = c("drug", "procedure"),
  firstTime = TRUE,
  attritionDomains = c("drug", "condition"),
  limit = "first",
  age = 18,
  gender = "male",
  exitStrategy = "observation_period_end_date"
)
}
}
