% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyCohortsToCreate.R
\name{writeListCohort}
\alias{writeListCohort}
\title{Write Cohort From List Expression}
\usage{
writeListCohort(cohort, path)
}
\arguments{
\item{cohort}{A list expression of the cohort.}

\item{path}{A character string specifying the path to save the cohort JSON file.}
}
\description{
This function writes a cohort list expression to a specified path in JSON format.
}
\examples{
\dontrun{
  writeListCohort(cohort, "path/to/cohort.json")
}
}
