% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptMappings.R
\name{standardFromSourceConceptIds}
\alias{standardFromSourceConceptIds}
\title{Get Mappings for Source Concept Ids}
\usage{
standardFromSourceConceptIds(cdm, sourceConceptIds)
}
\arguments{
\item{cdm}{cdm_reference via CDMConnector}

\item{sourceConceptIds}{A vector of source concept IDs.}
}
\value{
A vector of standard concept IDs.
}
\description{
This function retrieves standard concept IDs that map to given source concept IDs.
}
\examples{
\dontrun{
cdm <- tidyOhdsiRecipies::tidyCdmMock()
conceptIds <- conceptIdsFromSources(cdm, c("63020004901"), c("NDC"))
standardConcepts <- standardFromSourceConceptIds(cdm, conceptIds)
}
}
