% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyOhdsiRecipies-package.R
\docType{package}
\name{tidyOhdsiRecipies-package}
\alias{tidyOhdsiRecipies}
\alias{tidyOhdsiRecipies-package}
\title{tidyOhdsiRecipies: Methods for Working with OHDSI Tools}
\description{
Provides tools for working with OHDSI cohort definitions. This includes methods for managing cohorts, analyzing pathways, and supporting integration with the OHDSI ecosystem.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OdyOSG/tidyOhdsiRecipies}
  \item Report bugs at \url{https://github.com/OdyOSG/tidyOhdsiRecipies/issues}
}

}
\author{
\strong{Maintainer}: Alexander Alexeyuk \email{AlexanderAlexeyuk@gmail.com}

}
\keyword{internal}
