% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyCohortsToCreate.R
\name{writeListCohort}
\alias{writeListCohort}
\title{Write Cohort From List Expression}
\usage{
writeListCohort(cohort, cohortName, saveLocation)
}
\arguments{
\item{cohort}{A CirceR list expression of the cohort.}

\item{cohortName}{name of the cohort}

\item{saveLocation}{path to put cohort}
}
\description{
This function writes a cohort list expression to a specified path in JSON format.
}
\examples{
\dontrun{
writeListCohort(cohort, "any_cohort", "cohorts")
}
}
